refactor(persistence): decompose command_ledger god-functions into modules
Extract recover_expired_row (CCN 39) into recovery.rs with focused
helpers: load_recovery_evidence, validate_evidence_shape,
validate_clean_execution, reconcile_discovered_evidence,
commit_reconciled_terminal, reclaim_expired_lease.

Extract validate_terminal (CCN 22) into validate.rs with individual
invariant checks.

Extract finish_at SQL phases into finish.rs: guard_no_prior_evidence,
write_evidence_marker, write_audit_and_outbox, verify_written_counts,
commit_terminal_state. Movement-only — no behavior change.
